// src/app/woven/models/work-order.ts
export class WorkOrderDetail {
  wodetailid: any;
  workorderid: any;
  partsid: any;
  price: any;
  orderqty: any;
  customerpono: any;
  productionunitid: any;
  wastagepercentage: any;
  ordervalue: any;
  unitid: any;
  statusid: any;
  createdby: any;
  createddate: any;
  brandid: any;
  length: any;
  width: any;
  samplecolorid: any;
  finishtype: any;
  rcvstatus: any;
  adjrate: any;
  adjdate: any;
  adjby: any;
  adjamount: any;
  adjpriceid: any;
  adjqty: any;
  totalbreakdown: any;
  part: any;
  constructor() {
    this.orderqty = 0;
    this.adjamount = 0;
    this.customerpono = '';
    this.createdby = sessionStorage.getItem('empid');
    this.statusid = 0;
  }
}